requirements.txt
python-dotenv
audioop-lts
py-cord
loguru
captcha